LawRag/requirements.txt

python-dotenv
openai
langchain
langchain-openai
langchain-community
langchain-chroma
langgraph
chromadb

fastapi
uvicorn
psycopg2-binary
asyncpg
greenlet
pydantic
pydantic_core
SQLAlchemy
sqlmodel
starlette
email-validator